+++ internal/accrual/client.go
@@ -8,9 +8,9 @@ import (
 )
 
 type CalculationResponse struct {
-	Order      string `json:"order"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual,omitempty"`
+	Order      string  `json:"order"`
+	Status     string  `json:"status"`
+	Accrual    float64 `json:"accrual,omitempty"`
 	StatusCode int
 }
