+++ internal/entity/withdrawal.go
@@ -8,5 +8,5 @@ type Withdrawal struct {
 	OrderNumber int       `db:"order_number"`
 	ID          uint      `db:"id"`
 	UserID      uint      `db:"user_id"`
-	Sum         int       `db:"sum"`
+	Sum         float64   `db:"sum"`
 }
