+++ internal/handler/mocks/balance.go
@@ -35,10 +35,10 @@ func (m *MockUserBalanceStorage) EXPECT() *MockUserBalanceStorageMockRecorder {
 }
 
 // GetWithdrawalSumByUserID mocks base method.
-func (m *MockUserBalanceStorage) GetWithdrawalSumByUserID(arg0 context.Context, arg1 uint) (int, error) {
+func (m *MockUserBalanceStorage) GetWithdrawalSumByUserID(arg0 context.Context, arg1 uint) (float64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetWithdrawalSumByUserID", arg0, arg1)
-	ret0, _ := ret[0].(int)
+	ret0, _ := ret[0].(float64)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
